added LibraryModel, refactored BaseTableController

+++ src/main/java/controllers/book/tabs/BooksController.java
@@ -3,8 +3,9 @@ package controllers.book.tabs;
 import controllers.BaseTableController;
 import controllers.book.BookEditController;
 import controllers.student.StudentChooseController;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
@@ -17,18 +18,12 @@ import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 import model.Book;
-import service.book.BookService;
 
 import java.io.IOException;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.ResourceBundle;
 
-/**
- * Created by Vadim on 13.03.2017.
- */
-public class BooksController extends BaseTableController<Book> implements Initializable{
+public class BooksController extends BaseTableController<Book> implements Initializable {
     @FXML
     private TableView<Book> tvBooks;
     @FXML
@@ -39,13 +34,33 @@ public class BooksController extends BaseTableController<Book> implements Initia
     private TextField tfSearch;
 
     private static BooksController instance;
-    private ObservableList<Book> bookObservableList;
-    private BookService bookService = new BookService();
 
     public static BooksController getInstance() {
         return instance;
     }
 
+    @Override
+    public TableView<Book> getTableView() {
+        return tvBooks;
+    }
+
+    @Override
+    public TextField getTextFieldSearch() {
+        return tfSearch;
+    }
+
+    @Override
+    public void initTableData() {
+        observableList = FXCollections.observableArrayList(bookService.getAllBooks());
+        tcId.setCellValueFactory(new PropertyValueFactory<Book, Integer>("id"));
+        tcTitle.setCellValueFactory(new PropertyValueFactory<Book, String>("title"));
+        tcAuthor.setCellValueFactory(new PropertyValueFactory<Book, String>("author"));
+        tcEdition.setCellValueFactory(new PropertyValueFactory<Book, String>("edition"));
+        tcYearOfPublication.setCellValueFactory(new PropertyValueFactory<Book, String>("yearOfPublication"));
+        tcAvailability.setCellValueFactory(new PropertyValueFactory<Book, Boolean>("available"));
+        tvBooks.setItems(observableList);
+        tvBooks.setVisible(true);
+    }
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
@@ -59,19 +74,10 @@ public class BooksController extends BaseTableController<Book> implements Initia
     }
 
     private void setTextFieldFindBookListener() {
-        tfSearch.textProperty().addListener((observable, oldValue, newValue) -> {
-            String stringForSearch = tfSearch.getText();
-
-            if (stringForSearch.isEmpty()) {
-                tvBooks.setItems(bookObservableList);
-            } else {
-                List<Book> bookListByString = new ArrayList<>();
-                for (Book book : bookObservableList) {
-                    if (book.toStringForSearch().contains(stringForSearch))
-                        bookListByString.add(book);
-                }
-                ObservableList<Book> newList = FXCollections.observableArrayList(bookListByString);
-                tvBooks.setItems(newList);
+        tfSearch.textProperty().addListener(new ChangeListener<String>() {
+            @Override
+            public void changed(ObservableValue<? extends String> observable, String oldValue, String newValue) {
+                search();
             }
         });
     }
@@ -81,7 +87,7 @@ public class BooksController extends BaseTableController<Book> implements Initia
             if (getSelectionItem() != null && getSelectionItem().isAvailable()) {
                 Stage stage = new Stage();
                 Parent root = null;
-                FXMLLoader loader= new FXMLLoader(getClass().getClassLoader().getResource("fxml/student/student_choose_view.fxml"));
+                FXMLLoader loader = new FXMLLoader(getClass().getClassLoader().getResource("fxml/student/student_choose_view.fxml"));
                 try {
                     root = loader.load();
                 } catch (IOException e) {
@@ -98,9 +104,9 @@ public class BooksController extends BaseTableController<Book> implements Initia
 
     private void setButtonDeleteBookListener() {
         btnDeleteBook.setOnAction(event -> {
-            if (getSelectionItem() != null) {
+            if (getSelectionItem() != null && getSelectionItem().isAvailable()) {
                 bookService.deleteBook(getSelectionItem());
-                bookObservableList.remove(getSelectedId());
+                observableList.remove(getSelectedId());
             }
         });
     }
@@ -140,31 +146,5 @@ public class BooksController extends BaseTableController<Book> implements Initia
             stage.setScene(new Scene(root));
             stage.show();
         });
-    }
-
-    @Override
-    public void initTableData() {
-        bookObservableList = FXCollections.observableArrayList(bookService.getAllBooks());
-        tcId.setCellValueFactory(new PropertyValueFactory<Book, Integer>("id"));
-        tcTitle.setCellValueFactory(new PropertyValueFactory<Book, String>("title"));
-        tcAuthor.setCellValueFactory(new PropertyValueFactory<Book, String>("author"));
-        tcEdition.setCellValueFactory(new PropertyValueFactory<Book, String>("edition"));
-        tcYearOfPublication.setCellValueFactory(new PropertyValueFactory<Book, String>("yearOfPublication"));
-        tcAvailability.setCellValueFactory(new PropertyValueFactory<Book, Boolean>("available"));
-        tvBooks.setItems(bookObservableList);
-        tvBooks.setVisible(true);
-    }
-
-    @Override
-    public Book getSelectionItem() {
-        int id = getSelectedId();
-        if (id != -1) {
-            return bookObservableList.get(id);
-        } else return null;
-    }
-
-
-    private int getSelectedId() {
-        return tvBooks.getSelectionModel().getSelectedIndex();
     }
 }
